Stucturalisation of the components
